+++ NEWS.md
@@ -1,3 +1,6 @@
+
+# pkgbuild development version
+
 # pkgbuild 1.3.1
 
 * Accept Rtools40 for R 4.2, it works well, as long as the PATH includes