+++ NEWS.md
@@ -14,6 +14,9 @@
 * New `PKG_BUILD_COLOR_DIAGNOSTICS` environment variable to opt out from
   colored compiler output (#141).
 
+* pkgbuild now works with a full XCode installation if the XCode Command
+  Line Tools are not installed, on macOS, in RStudio (#103).
+
 # pkgbuild 1.3.1
 
 * Accept Rtools40 for R 4.2, it works well, as long as the PATH includes