+++ NEWS.md
@@ -1,5 +1,4 @@
-
-# pkgbuild development version
+# pkgbuild 1.4.0
 
 * pkgbuild can now avoid copying large package directories when building a
   source package. See the `PKG_BUILD_COPY_METHOD` enviroment variable in